Extract websocket code into own module
